Use case-insensitive search for TGA material files
After attempting to open the TGA file for a given
3ds material directly, fall back to trying to
locate it using a case-insensitive search for the
file name in the directory where the 3ds file was
found.  This avoids the need to have awkwardly-
named SOMETHING.TGA files when using 3ds files
generated on DOS/Windows/Mac which have case-
insensitive filenames.